refactor(msal-browser): Isolate Native Auth V2 auth_flow files under v2/
Move CustomAuthV2FlowScenario and CustomAuthV2Result from core/auth_flow/
into core/auth_flow/v2/ so all V2 auth_flow code lives under a single v2/
subtree. This keeps V2 fully separable from V1 and eases future V1
deprecation (removal becomes "delete everything not under a v2/ folder").
Import paths updated across importers; no behavior change.

Copilot-Session: 73173219-90a6-4d29-8124-d1245009f8ab
