+++ lib/msal-browser/src/custom_auth/core/auth_flow/v2/CustomAuthV2FlowScenario.ts
