+++ lib/msal-browser/src/custom_auth/core/auth_flow/v2/result/ResetPasswordStartV2Result.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { CustomAuthV2Result } from "../../CustomAuthV2Result.js";
+import { CustomAuthV2Result } from "../CustomAuthV2Result.js";
 import type { ResetPasswordStartError } from "../error/ResetPasswordStartError.js";
 import type { AuthenticationMethodSelectionRequiredState } from "../state/AuthenticationMethodSelectionRequiredState.js";
 import type { FailedState } from "../state/FailedState.js";
