+++ lib/msal-browser/src/custom_auth/core/interaction_client/v2/V2FlowContinuationState.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { CustomAuthV2FlowScenario } from "../../auth_flow/CustomAuthV2FlowScenario.js";
+import { CustomAuthV2FlowScenario } from "../../auth_flow/v2/CustomAuthV2FlowScenario.js";
 
 /*
  * Server-provided `_links` hrefs for the next step(s) of a flow, resolved from the previous HAL
