+++ lib/msal-browser/src/custom_auth/index.ts
@@ -252,10 +252,10 @@ export {
 } from "./CustomAuthV2ActionInputs.js";
 
 // Native Auth V2 Result
-export { CustomAuthV2Result } from "./core/auth_flow/CustomAuthV2Result.js";
+export { CustomAuthV2Result } from "./core/auth_flow/v2/CustomAuthV2Result.js";
 
 // Native Auth V2 Flow Scenario
-export { CustomAuthV2FlowScenario } from "./core/auth_flow/CustomAuthV2FlowScenario.js";
+export { CustomAuthV2FlowScenario } from "./core/auth_flow/v2/CustomAuthV2FlowScenario.js";
 
 // Native Auth V2 Authentication Method
 export { AuthenticationMethodV2 } from "./core/auth_flow/v2/AuthenticationMethodV2.js";
